// src/containers/planning/enum.ts
export enum PlanningActionTypes {
	SET_PLANNING = 'SET_PLANNING',
	SORT_PLANNING_TASKS = 'SORT_PLANNING_TASKS',
	DELETE_PLANNING_TASK = 'DELETE_PLANNING_TASK',
	UPDATE_PLANNING_TASK = 'UPDATE_PLANNING_TASK',
	CHANGE_PLANNING_TASK_STATUS = 'CHANGE_PLANNING_TASK_STATUS'
}

export enum EventActionTypes {
	SET_EVENT_PARAMETER = 'SET_EVENT_PARAMETER',
	ADD_EVENT = 'ADD_EVENT',
	SET_EVENTS = 'SET_EVENTS',
	SET_EVENT = 'SET_EVENT',
	UPDATE_EVENT = 'UPDATE_EVENT',
	SET_VAT = 'SET_VAT',
	DELETE_EVENT = 'DELETE_EVENT'
}